Move XML generation into private submodules and remove top-level Coordinator access
