+++ pyoozie/__init__.py
@@ -1,13 +1,13 @@
 # Copyright (c) 2017 "Shopify inc." All rights reserved.
 # Use of this source code is governed by a MIT-style license that can be found in the LICENSE file.
 
-from pyoozie.coordinator import Coordinator, ExecutionOrder
-from pyoozie.tags import Parameters, Configuration, Credentials, Shell, SubWorkflow, GlobalConfiguration, Email
-from pyoozie.builder import WorkflowBuilder, CoordinatorBuilder
 from pyoozie.api._model import parse_coordinator_id, parse_workflow_id, ArtifactType, \
     CoordinatorStatus, CoordinatorActionStatus, WorkflowStatus, WorkflowActionStatus
 from pyoozie.api._client import OozieClient
 from pyoozie._exceptions import OozieException
+from pyoozie.xml._coordinator import ExecutionOrder
+from pyoozie.xml._tags import Parameters, Configuration, Credentials, Shell, SubWorkflow, GlobalConfiguration, Email
+from pyoozie.xml._builder import WorkflowBuilder, CoordinatorBuilder
 
 __version__ = '0.0.0'
