+++ tests/pyoozie/xml/test_builder.py
@@ -9,7 +9,7 @@ from datetime import datetime
 import pytest
 
 from pyoozie import WorkflowBuilder, CoordinatorBuilder, Shell, Email, ExecutionOrder
-from pyoozie.builder import _workflow_submission_xml, _coordinator_submission_xml
+from pyoozie.xml._builder import _workflow_submission_xml, _coordinator_submission_xml
 from tests.utils import xml_to_dict_unordered
